[!!!][TASK] Adjust ViewHelperInterface to cover current API
In the past, some ViewHelper API methods that were used by Fluid
internally were only part of `AbstractViewHelper`, which lead to
either `method_exists()` checks or phpstan warnings in the baseline.

With this change, the `ViewHelperInterface` becomes the internal
API definition for ViewHelpers, only to be used by Fluid internally.
Thus, missing API methods are added to the interface, while
user-facing API only used by `AbstractViewHelper` has been
removed.

In the future, we might consider adding a separate interface for the
used-facing API of ViewHelpers. For now we highly recommend
to keep using `AbstractViewHelper` as the basis for all custom
ViewHelper implementations.