[TASK] Remove renderStatic() from condition ViewHelpers (#989)
Since `renderStatic()` is now deprecated, `AbstractConditionViewHelper` is
adjusted accordingly. Template compilation steps are left as-is except for the
ViewHelper call itself, which now also uses `ViewHelperInvoker`. The
functionality of `renderStatic()` is now integrated into the separate render
methods. Also, special handling for the internal closure arguments is added
to the class.