[TASK] Prevent template compilation in inheritance test (#999)
This is an ugly workaround to prevent the interfering runtime
cache until #975 is resolved. Currently, subsequent test cases
are successful even if the rendering is modified to not inherit
ViewHelper namespaces to sub templates because they
re-use the "in-memory" template classes from previous tests.
As there is currently no way to configure this, we need to use
the ugly f:cache.disable ViewHelper in the affected template
to get robust testing results.