[TASK] Catch errors in schema generation and collect them (#1008)
There are cases when errors occur on iterating through the installed view helpers. The generation of the XSDs then break hard with an exception. Currently, this occurs with the TYPO3 extension "schema".

Now, these errors are catched and collected. A calling script can get and display these errors. This way, the generation of XSDs is not cancelled, but proceed without the erroneous view helpers.

This is especially useful for the TYPO3 command "fluid:schema:generate".

Related: #1005