[TASK] Avoid some useless instanceof tests
Some tests verify if view helper classes are an
instance of an abstract to see if the VH "can
be instantiated". Those are useless and
instantiation of those VH's is indirectly tested
by following tests that render a template using
the VH.

The instanceof tests add no value and are removed,
testing for the abstracts is useless anyway and
the only contract that must be adhered to is
ViewHelperInterface, which is tested by the
following tests.