[BUGFIX] Ensure consistent handling of else-if ViewHelpers
A f:if ViewHelper that contains else-if ViewHelpers and also sets the else
attribute behaved differently for cached and uncached templates. With
this change, uncached templates now behave just like cached templates.

This can be validated by temporarily commenting out the second test
execution code in IfThenElseViewHelperTest::render(). This testing setup
issue should also be fixed in the future, but this is out of scope for the
specific issue.