[BUGFIX] Consistent return types for condition-based ViewHelpers
In most cases, condition-based ViewHelpers already return an empty string
as default value, for example if the verdict is false and else is not set. However,
there are a few cases where `null` would be returned instead, which would
in the template implicitly converted to an empty string in the vast majority of
cases anyway.

This change makes sure that all closures or direct ViewHelper node evaluations
default to an empty string if `null` is returned, making the behavior consistent
to other code paths where an empty string is returned explicitly.