[FEATURE] ViewHelpers without then/else return verdict (#1031)
ViewHelpers that don't specify any then or else clause in any of the accepted
syntax variants now directly return the result of the condition.

Previously, this syntax was not supported and a no-op resulting in an empty
string. Thus, this is not considered a breaking change, but an enhancement
of the existing behavior.