[TASK] Add SchemaGenerator evaluation for "@see" (#1047)
The phpDoc annotation "@see" contains valuable information
for the XSD schema generation for future reading.

Even though some IDEs like PHPStorm will not expose these links
as being clickable, they can be copied+pasted from a tooltip
in the IDE, and opened in the browser (instead of needing to
find and open a PHP classfile to get to that link).

Some IDEs (eclipse?) seem to be able to parse a HTML subset
for annotations, so maybe this will find future adoption
(or markdown rendering might get available). Since also
neither ReST nor Markdown is parsed in the XSD element
`xsd:documentation` by PHPStorm either.... plaintext it is!