[BUGFIX] Ignore f:then/f:else outside condition context
Before this change, both `f:then` and `f:else` could be used outside of
a condition tag and would then always output the content, regardless
of any condition. Now, nothing is outputted.

In the future, we might consider throwing an exception here to alert the
user about invalid Fluid syntax. However, this would be a breaking change.

Resolves: #1020