[BUGFIX] Differentiate between null and unset then/else condition branches (#1051)
With Fluid 4.1 (in 96099a6ba60bcb6a73187d289a781903a05882e3) it became possible to
use `f:if` and other condition-based ViewHelpers without `then` or `else`, which
would return the result of the condition as boolean. However, the current
implementation of this feature didn't consider the difference between a ViewHelper
without any `then` or `else` and one with `then` that returns `null`. These two
examples would both return the boolean:

```
{f:if(condition: myCondition, then: myUndefinedVariable)}
{f:if(condition: myCondition)}
```

With this change, the two examples are treated differently. Instead of checking
the arguments of the ViewHelper (which are already merged with the defined
default values), the ViewHelperNode is used to find out if the argument has
actually been specified. This applies both for uncached rendering and compilation,
which is used for cached templates.

Since the implementation in `AbstractConditionViewHelper` became more and more
convoluted, both the rendering and the compilation step have been refactored.
Also, more test cases are added, which make sure that the issues of the current
implementation are still addressed with the refactored code.

Resolves: #1049