[BUGFIX] Address regression in condition VH with else attribute
The recent changes on condition-based ViewHelpers introduced a regression
if a ViewHelper specifies "else" as an attribute, while "then" is specified
implicitly via the tag's content:

```
<f:if condition="{someVar}" else="{someStringVar}">It works!</f:if>
```

In that case, the "then" portion ("It works!") is ignored due to a bug
in cache compilation. This has been addressed with this patch and the
missing test cases have been added to the test suite.