[FEATURE] StrictArgumentProcessor
The `StrictArgumentProcessor` is a stricter variant to the newly
introduced `LenientArgumentProcessor`. This patch only contributes
the implementation of the argument processor itself, without
actually using it in the code.

The plan is to use the new strict processor for upcoming new
features and to switch the existing ViewHelper implementation
to the stricter variant with Fluid v5. There probably will also be
an API to provide alternative implementations to Fluid in the
future once the `StrictArgumentProcessor` is used across
the whole code.