[FEATURE] Introduce ArgumentProcessor API (#1062)
Fluid allows ViewHelpers to specify their API by using
`AbstractViewHelper::registerArgument()`, which is usually called in
`initializeArguments()` of the ViewHelper. Internally, `ArgumentDefinition`
is used to represent the provided argument information and constraints.
Among those is the ability to specify the type of an argument as well
as if the argument should be a required argument.

While the required status is already validated at the parser level, the
argument types are validated by the ViewHelper implementation. However,
in most cases, Fluid's built-in validation method by `AbstractViewHelper`
is used. Unfortunately, this implementation has many flaws:

* it only covers a subset of possible types (e. g. it covers "boolean", but
  not "bool")
* It doesn't ensure the correct argument type for scalar arguments (e. g.
  you might receive an integer in your ViewHelper implementation, even
  if the argument has been defined as "string")
* There are several small inconsistencies

In order to fix this in a future Fluid version without introducing too many
hard-to-debug breaking changes, a new API is introduced which extracts
validation functionality from the `AbstractViewHelper` and allows
alternative implementations in the future. It also adds the possibility
to pre-process arguments before their validity is checked, which will
allow implicit type conversions in the future (e. g. convert "int" to "string").

As a first step, the existing helper methods `isValidType()` and
`getFirstElementOfNonEmpty()` are deprecated and extracted into
a `LenientArgumentProcessor`, which mimicks the current validation
behavior. Like the current implementation, it doesn't modify the argument
types, but rather passes them along unmodified.

The new API is then called for the pre-processing step in
`ViewHelperInvoker` as well as the validation step in `AbstractViewHelper`.
In this first implementation, it isn't possible yet to replace the default
implementation with an alternative.