[FEATURE] Introduce ViewHelperResolver delegates
It has always been possible to set a custom `ViewHelperResolver` for
Fluid in the rendering context. One major use case was (and still is)
to use a custom ViewHelper factory or dependency injection. However,
because there can only be one `ViewHelperResolver`, it becomes quite
hard to extend/adjust Fluid's ViewHelper resolving (= which class should
handle which ViewHelper call in a template) to special needs.

With the introduction of ViewHelperResolver delegates, the main
ViewHelperResolver can delegate the actual resolving within a
namespace to a custom implementation. This means that the logic
behind `<f:format.trim>` and `<my:custom>` can now be
different: Class names can have a different structure, and it's no longer
strictly necessary to have one ViewHelper class per ViewHelper. The
reasoning behind this new API is the ongoing work on the "reusable
components" feature, but this might also offer other opportunities
in the future, like bridges to other templating engines or more dynamic
ViewHelper resolving by third-party packages.

This patch introduces a new `ViewHelperResolverDelegateInterface`,
which needs to be implemented by custom delegates. The current
resolving logic is moved from `ViewHelperResolver` to a new
`ViewHelperCollection` class, which implements the new interface.

The `ViewHelperResolver` is now responsible for creating and
managing these delegate instances and to find the appropriate
delegate for a specified PHP namespace. Delegates implement
the interface method `resolveViewHelperClassName()`, which
either returns the PHP class name for the supplied ViewHelper
name or throws `UnresolvableViewHelperException` if the
delegate can't find an appropriate ViewHelper class. If no resolver
is able to find a class name, the `ViewHelperResolver` throws
the same exception as before this change.

The delegate for a namespace is determined by first checking
for the existence of the PHP class name the Fluid namespace refers
to. Only if no class exists, the previous logic applies.

Example:

```xml
{namespace a=Vendor\PackageA\ViewHelpers}
{namespace b=Vendor\PackageB\ViewHelperCollection}

<a:custom />
<b:custom />
```

```php
<?php

namespace Vendor\PackageB;

use TYPO3Fluid\Fluid\Core\ViewHelper\ViewHelperResolverDelegateInterface;

final readonly class ViewHelperCollection implements ViewHelperResolverDelegateInterface
{
    public function resolveViewHelperClassName(string $name): string
    {
        // custom resolver logic
    }

    public function getNamespace(): string
    {
        // string representation of the delegate, usually its
        // fully qualified name
        return self::class;
    }
}
```

Namespace `a` would follow the same default rules as before: The
ViewHelper `<a:custom>` would resolve to
`Vendor\PackageA\ViewHelpers\CustomViewHelper`.

Namespace `b` however would follow the custom rules implemented
in the delegate: `Vendor\PackageB\ViewHelperCollection` would be
instantiated by `ViewHelperResolver` and
`resolveViewHelperClassName('custom')` would be called, which
either would return a PHP class name or would throw the exception
as described above.

The patch also makes sure that already instantiated delegates can
be passed directly to the `ViewHelperResolver`.