[TASK] Make namespace inheritance tests more consistent (#1076)
This expands our test cases for the more obscure feature
of namespace inheritance, which is due to be deprecated
and removed. It allowed partials to access ViewHelpers
by "reusing" the import from their parent template.

The tests now cover more use cases and should also be
more reliable with cached templates.