[TASK] Deprecate ViewHelper namespace inheritance
Fluid v5 will no longer support inheritance of ViewHelper namespaces
from a template to its layout or partials. Namespace must either be
added globally (by using the public `ViewHelperResolver` API) or
by using the `{namespace x=…}` or `xmlns:x="…"` syntax in the
template file that uses the ViewHelper.

This patch prepares the removal of this feature by implementing
separate namespace collections for global, local and inherited
ViewHelper namespaces. Depending on the context, ViewHelpers
will be added to the appropriate collection. When a ViewHelper
is called, the different collections will be checked for a matching
namespace. Also, contexts get their own `ViewHelperResolver`
instances to isolate them properly.

If a namespace is only defined in a parent template, but not in
the template itself or in the global collection, a deprecation
will be emitted, which allows to prepare existing templates
before switching to Fluid v5 later on. To resolve a deprecation,
all ViewHelper namespaces that are used within a template
need to be either defined globally (like f:*) or directly in the
template file.

With Fluid v5, the backwards compatibility layer for inherited
namespaces will be removed, which will be a breaking change.
This change is a hard requirement for a more reasonable
cache warmup feature for Fluid templates.