[TASK] Provide identifier early to ParsingState (#1081)
Currently, the `ParsingState` only gets its (optional) identifier after
the whole parsing process, even though it's already available at
the beginning. With this change, the identifier is set early (and
to an empty string as fallback).

This allows us to provide more context in deprecation messages
which will be added in follow-up patches.