[FEATURE] Reusable components
This major feature introduces the possibility to map Fluid templates to
ViewHelper tags. This makes it easy to "invent your own elements", also
known as defining components. Together with the recently introduced
`<f:argument>` ViewHelper, this creates a powerful alternative to partials,
which have one major shortcoming: Partials can only be used if their path
is added to the list of `partialRootPaths` in the current template context.

With the new components mechanism, this is no longer necessary. Instead,
similar to ViewHelpers, collections of components can be defined and
imported into arbitrary templates by using either the `{namespace …}`
syntax or its `xmlns:` equivalent. For that to work, a PHP class needs
to be provided, which specifies where the component files for the
namespace are located:

```php
<?php

namespace Vendor\Package\Components;

use TYPO3Fluid\Fluid\Core\Component\AbstractComponentCollection;
use TYPO3Fluid\Fluid\View\TemplatePaths;

final class MyComponentCollection extends AbstractComponentCollection
{
    public function getTemplatePaths(): TemplatePaths
    {
        $templatePaths = new TemplatePaths();
        $templatePaths->setTemplateRootPaths([
            'path/to/Components/',
        ]);
        return $templatePaths;
    }
}
```

Inside our `Components/` folder, component templates can be placed.
By default, each component has its own folder, in which its html
template is placed.

`path/to/Components/Button/Button.html`:

```xml
<f:argument name="label" type="string" />

<button class="myButton">{label}</button>
```

This button component can then be rendered by arbitrary Fluid
templates, without any further configuration, by importing the
namespace of the component collection class:

```xml
{namespace my=Vendor\Package\Components\MyComponentCollection}

<my:button label="Label of my button" />
```

The result would be:

```xml
<button class="myButton">Label of my button</button>
```

Also supported is a way to render the child content of a component
tag by using the new `<f:slot>` ViewHelper:

`path/to/Components/Quote/Quote.html`:

```xml
<f:argument name="author" type="string" />

<blockquote>
    <f:slot />
</blockquote>
<p>—{author}</p>
```

Again, this can be rendered in any Fluid template:

```xml
{namespace my=Vendor\Package\Components\MyComponentCollection}

<my:quote author="Jane Doe">
  <p>This is the quote</p>
</my:quote>
```

There are some noteworthy details about the implementation
and the API:

The ComponentAdapter is an adapter class that translates
between the ViewHelper API and the newly created components
API. It is marked `internal` and is not intended to be replaced.

In addition to the previously introduced
`ViewHelperResolverDelegateInterface`, there are now three more
interfaces that are specific to components: The
`ComponentDefinitionProviderInterface`, the
`ComponentRendererInterface` and the
`ComponentTemplateResolverInterface`.

The `ComponentDefinitionProviderInterface` defines how abstract
component definitions (described by the also introduced
`ComponentDefinition` data structure) can be made available to
Fluid. This is the interface the already mentioned `ComponentAdapter`
needs to bridge between ViewHelpers and components. For
Fluid-based components, this interface is not relevant because there
is already an implementation in `AbstractComponentCollection`
(which was used in the example above). Same goes for the
`ComponentRendererInterface`, which is already implemented for
rendering Fluid templates in the default `ComponentRenderer`.
These interfaces can however be used to bridge other *things*
into Fluid, like other Templating engines.

The third interface `ComponentTemplateResolverInterface` is an
internal interface that is specific to Fluid's component
implementation (meaning: Using Fluid templates as components)
in `AbstractComponentCollection`.

From a user perspective, the `AbstractComponentCollection`
provides a default base implementation for most use cases. A
minimal implementation (as can be seen in the example above)
only needs to provide `getTemplatePaths()`. However, it is also
possible to alter the path structure by overriding
`resolveTemplateName()`, to add additional variables to component
templates by overriding `getAdditionalVariables()` or to allow
arbitrary arguments for components by overriding
`additionalArgumentsAllowed()`.

An unfortunate shortcoming of the current `TemplateParser`
is that it needs to resolve and validate all ViewHelpers within
a template at parse time, which can lead to various unintended
consequences in the context of components. For the time being,
`TemplateStructureViewHelperResolver` is necessary to inspect
component templates without triggering a long chain of
unrelated parsing tasks. This has been documented in
`AbstractComponentCollection->getComponentDefinition()`.

Detailed RST documentation will be added in follow-up commits.