[BUGFIX] Attempt to cast boolean arguments
The recently introduced `StrictArgumentProcessor` didn't process
boolean arguments because it relied on the `BooleanParser` being
performed during template parsing. However, this is not the case
for templates being rendered directly from PHP:

```php
$myView->assign('booleanArgument', 1);
$myView->render('myTemplate');
```

This change assures that the `1` will be interpreted as valid boolean,
which is consistent with the other scalar types.