[TASK] Clarify parsing state
Currently, the state handling in the `TemplateParser` is quite confusing.
The actual result of the parser is the `ParsingState`, which will later be
used to compile the template for the cache, but can also be rendered
directly. However, there isn't necessarily only one `ParsingState` object
per template: In some syntax variants, a "sub template" with its own
`ParsingState` and `RootNode` is created. Once the sub parsing is
complete, the `RootNode` will be extracted from the inner
`ParsingState` and will be appended to the outer `ParsingState`.
However, this is not the case for other state in that object.

As a first step, the outer `ParsingState` is created at the first level of
the parsing process and is handed down via method arguments.
The internal template now explicitly creates a temporary `ParsingState`
with the identifier `null`. The getter method is renamed to
`createParsingState()`, because that's what the method does.