[TASK] Add Contribution Guidelines (#1120)
This commit adds an initial set of contribution
guidelines to the Fluid project. These guidelines
enable new contributors to see a quick summary of
required steps to make a contribution and information
on the local development setup and tooling that
makes it possible to work on Fluid.

It has been chosen to add these guidelines as a
Markdown file in `.github/CONTRIBUTING.md` to
support the GitHub feature that makes new contributors
aware of existing Guidelines when creating a Pull
Request.

Related: #1116