[FEATURE] Named Slots (#1118)
This patch enables Fluid's components to accept multiple slots. Each
slot is identified by a `name` argument. The corresponding
`<f:fragment>` ViewHelper can then be used to fill the slots of a
component.

The available slots of a component are now also exposed properly
via the `ComponentDefinition` DTO. Unfortunately, the `TemplateParser`
fails to propagate available slots of the current templates consistently to
the `ParsingState`, which currently leads to ugly state merging within the
parser. Hopefully, this issue can be addressed in the future to remove this
workaround, which has been documented in the code.

For this first implementation, we limit the usage if the new
`<f:fragment>` ViewHelper to the context of components. However, we
might expand this ViewHelper to other areas in Fluid, such as a way to
partially extend/overwrite templates.