[TASK] Clarify code comments related to escaping (#1151) (#1152)
A code search both for "HtmlspecialcharsViewHelper" and
"htmlspecialchars" confirmed that the ViewHelper is no longer used
internally for escaping. Instead, the `EscapingNode` executes
`htmlspecialchars()` directly on the string, both in uncached and
cached context.

The `@todo` in `HtmlspecialcharsViewHelper` is removed without
changing the return type though, because this would require
a breaking change to the ViewHelper's internal behavior.