[TASK] Disallow argument definitions with a default value while also being required (#1157)
If a default value is supplied for a ViewHelper or component argument, but the
argument is defined as required, an exception is now thrown to make the user
aware that the default value will not be considered.

This change will not be backported to 4.x because it might break existing
ViewHelpers or components if they are not configured properly.