[TASK] Deprecate unknown VH namespace syntax
There are two main syntax variants to import a ViewHelper namespace
in a template:

```
{namespace my=Vendor\MyPackage\ViewHelpers}

<html xmlns:my="http://typo3.org/ns/Vendor/MyPackage/ViewHelpers" data-namespace-typo3-fluid="true">
```

The second variant is mainly used/preferred because it provides
autocompletion for ViewHelpers in supporting IDEs. However, a
third variant is implemented, which mixes the two variants:

```
<html xmlns:my="Vendor\MyPackage\ViewHelpers" data-namespace-typo3-fluid="true">
```

This doesn't make much sense because it neither provides
autocompletion nor is it valid XML syntax because it's
not a valid URL.

This patch deprecates that syntax variant, which will no longer
work in Fluid v5.

Resolves: #1019