[DOCS] Improve TYPO3-related sections of components docs (#1188)
With TYPO3 v14, there will be a deeper integration of the new
components feature into TYPO3. At that point, TYPO3-specific
documentation of components can be added to the official TYPO3
documentation.

For now, we try to cover TYPO3-specific parts in the Fluid
documentation by adding tips in the relevant places.