[!!!][TASK] Strict argument processing for ViewHelpers
Fluid v4.2.0 introduced a new API to process component and ViewHelper
arguments (see #1062). To avoid breaking changes in Fluid v4 with
how ViewHelpers handle their arguments, a `LenientArgumentProcessor`
has been introduced which has been used exclusively for ViewHelpers.
Components have used the new `StrictArgumentProcessor` from the
beginning.

With this breaking change, this is now unified: `StrictArgumentProcessor`
is now used for ViewHelper arguments as well. This allows us to define
an API in a follow-up patch that allows developers to define their custom
argument processing and validation logic or extend/wrap the existing
logic by Fluid.

The `StrictArgumentProcessor` deals with argument types in a stricter
way: If the supplied argument value doesn't meet the type requirement,
an exception is thrown. For scalar types, automatic type conversion
is happening in a processing step to avoid errors that can be easily
solved by Fluid itself (like "string is required, integer given").

There might be consequences for custom ViewHelpers, however in
most cases this change will make a lot of custom validation code
that was necessary until Fluid v5 obsolete.