[BUGFIX] Call proper API to obtain ViewHelper API
Currently, the ViewHelper validation API calls the internal ViewHelper
API to obtain its argument definitions for validation. However, the
proper way is to use the `ViewHelperResolver`, since it is possible
to redefine certain argument definitions there.

This is demonstrated and tested by the `CustomViewHelperResolver`
in the examples. However, this example only worked until now because
of the lenient validation implementation of ViewHelpers.