[TASK] Avoid context-dependent prefixes for cache files (#1205)
In order to implement a reasonable cache warmup, the template
cache needs to be independent of the context where the template
will be used later in the project. For cache files, this means that
they cannot limit how a template will be used later on
(partial, layout or template).

This patch removes the context-dependent prefixes for cache
file names to achieve this.