[!!!][FEATURE] Ability to replace argument processor
This patch introduces new API methods on the RenderingContext to
obtain and replace Fluid's argument processor implementation. By
default, the StrictArgumentProcessor is used, but the
ArgumentProcessorInterface exists to provide an alternative
implementation. However, it might be a good idea wrap custom
implementations around the StrictArgumentProcessor by using
the decorator design pattern to ensure valid input types
to ViewHelpers and components.

This is a breaking feature because of the interface change.

Resolves: #1105